Expand filters to work beyond simple aggregations
Before this filters were built from an ES aggregation, so the name of
the aggregation needed to match a field in ES. Aggregations like 'room'
and 'artist' worked, but 'gist' and 'on view' didn't because those
aggregations aren't `terms` aggregations.

To get a 'fancy' aggregation to be filter-able, pass a value in the
`customFilters` object that matches the name of the aggregation.