WIP: click to select does a 'peek' at facets
The idea: when a facet is interesting, click on it to open a preview
popup with more of that content. So on seeing 'Three Bathers', a click on
'Henry Matisse' will activate a 'peek' for other art with `artist:"Henry
Matisse"`.

Or I want to see all the other stuff in G321 along with the thing I'm
looking at. Click to select, open a popup, bam.

From there I could continue on to that search, dive straight into an
individual object, or close the peek and go about my business.