scroll fixed right-hand boxes
Artificially scroll the fixed sticky box on the right along with the
results on the left by adjusting the right box `top` to an increasing
negative number.

this is a bit sketchy. And it breaks when the left column is shorter
than an object stuck to the right. But it'll work for now?