improvements from Diane
quoting…

Artist IDs: We don't use "Unknown artist"; instead, we state the
tribe/country of origin. So the "artist" name for Jade Mountain becomes
China.

Dates/ranges: For brevity and style, we combine numbers in a century;
for instance, 1856 - 1858 becomes 1856–58, with an en dash, no spaces.

Caps: We also lowercase material names unless it's proprietary. So, for
Jade Mountain, the Light would need to become light green jade.

Titles: We italicize titles that were given by the artist or are
descriptive. In other words, Three Bathers vs. Hunting Coat.