fix bug that capped size of results
Because of the order the objects were included in the query used to
build the next link, any `size` value that was already set overrode the
intended next size.

You could click the more link once to go from 100 to 300, but clicking
'load more' a second time, the 500 was replaced by the existing 300 and
it never actually loaded the next set of results.  Putting the new
`size` after the existing `props.query` keeps everything else in the old
`query` while writing the new intended `size` so users can scroll and
scroll.