update broken icon links
This swaps in the material email icon and changes the sharing icons to
black on white.

material icons doesn't have facebook and twitter links, so the
sharing icons are linking to an external icon set with brand logos[1],
which had changed it's URL structure and broken our links.

*Simple Icons* repo also removed their general email icon, and changed
their icons to have black foregrounds.

[1] https://simpleicons.org/