guess audio content when no artwork is associated
Audio Stops are a set of 3 digit numbers, so if one of them is searched
and we can find an audio tour file that matches, show exclusively the
matching file if there's not a hit coming from ES that matches. (If
there is an audio hit in ES use that to render a fuller view of the
audio with the artwork info included).