fix scroll adjustment problems
The side-by-side search results viewer modifies the scroll of a
fixed-position artwork preview. The code for adjusting that scroll was
causing problems - mainly sometimes it wouldn't reset from `fixed` to
normal position, resulting in the artwork preview being stuck to the
very top of the window, and disappearing behind the search bar.

This refactors the scrolling code and fixes that problem.