slightly improve 'back' button on artwork page
The problem is we don't necessarily know what 'back' goes to - it might
be something outside of out website. (For instance a blank page, if the
user directly pasted in a link to the artwork…) This is a
web/html/browser/security restriction.

This changes it so if there is no history, the back button will instead
say 'home' and link to the homepage. More advanced routing could be
used, so that search history is saved, which would make it easy to do
the back button.