Expand `art.room` to 'On View {art.room}'
The raw TMS text, `room = 'G…'` is inscrutable and people don't always
get what that means. Use better language to convey a clearer meaning.

This is a bit tricky because of how the search results get highlighted:
with a search for G220, ES helps us call out where/how the results match
the search query by providing a highlight. With this patch if there's a
highlight on the gallery field we lose the 'on view' language. But I
think that's OK because if a user is searching directly for a gallery
number they probably get that G123 means 'On View'