get pages working in the wayback machine
take 1: patch react-router to re-write the page URL behind the scenes to
plaster over the fact that the App is running with a URL structure
modified by archive.org. Also add a few important links as
`rel=prefetch` so it wraps up all the data it needs to continue to work.