Start prototyping 'Inspired'/Foot in the Door UX
Focused on showing non-Mia artworks within the collections site
framework. For Inspired By Mia - they are user-submitted artworks
inspired by works in the collection. Connect them as related content and
design a special view that replaces the Mia artwork image in the search
results with the user submitted artwork image.

When foot in the door data is ready it will use the same design, but
those works are standalong, not related to Mia works so it will work a
tiny bit differently.

This is also being used as the first step on the process of redesigning
the collections site