pre-empt exhibition with a 'closed' message
when the exhibition 'closes'. Foot in the Door is over, and that needs
to be communicated clearly due to the works not being owned by the
museum.

Cover the landing pages with a 'page blocking' overlay that includes a
message with details about how the show has ended and why it's no longer
accessible online, while also leaving individual artworks up for
historical reference, but including the same 'its over' messaging.