feat: enable/disable search per-exhibition
This is a bit of a fudge to make time to decide how to search within an
exhibition.

Foot in the Door was easy - there are tons of things to search for and
they are in a separate ES index.

For an exhibition like Todd Webb, `search.artsmia.org` doesn't have a
way to search just the works in the exhibition yet.