feat: slice sibling artworks in the grid
Show the ~30 artworks directly preceding and following below the main
artwork on /exhibitions/:id/art/:id

This is instead of only showing the first ~30, and instead of showing
every single work, which gets overwhelming in a show with hundreds of
items.

IDEA: build in the 'show more' button here? Make the images a
lazy-loading, one-line scrolls so they can all exist below without
potentially loading hundreds of MBs?