fix: wait for IIIF generation to load deepzoom
Previously the library called `tiles.dx.artsmia.org` to generate IIIF
tiles if they didn't exist and `iiif.dx.artsmia.org` for the canonical
`info.json` independently.

This meant that if the tiles didn't exist, the code would recieve a 404
on the info.json and hang on the 'loading' message.

Calling for both resources in sequence means that the info.json will be
generated on both `tiles.` and `iiif.` and as soon as the IIIF tiles are
generated, the image will load in the browser.