[MS25-68] Avoid React rendering bug by avoiding nested anchor elements
Nested anchor elements are "fixed" by browsers, and this reorganization
confuses React's DOM renderer, breaking display. Hence, we try to pass
a prop into Markdown uses that are likely to display inside an anchor,
and Markdown strips A tags.

Background: URL route `/search/*/filters/room:"G371"` returns artwork 1244
as the first element and this includes an `image_copyright` property with
a URL. Because the property is rendered inside an anchor, this resulted
in Markdown creating a nested anchor.