Don't force exit to the Info site
Clicking the logo *should* just open the menu. But if the click event is
allowed to propagate in the DOM, the link in the menu will end up clicked,
too, causing the user to be unexpectedly navigated to the Info site instead
of having a chance to view the expanded menu.

This prevents that unintentional click from occurring.